+++ changelog.md
@@ -1,3 +1,8 @@
+latest
+ - Nudge Crossings selected action, nudges all zero crossings off from zero.
+ - Pad with Zero selected action, adds a zero sample to the beginning and end of each file.
+ - Shift selected action, moves the sample start position over 50%.
+
 2024.05.28 1.4.15
  - Amplitude threshold condense selected action, enter an upper and lower amplitude value (between 0 and 1), the sample will be modified to only include amplitudes within the given range, gaps between the audio is removed condensing the sample size.
  - Ctrl + Click on Filename column header to sort the list by numeric filename content only.